\name{trans-package}
\alias{trans-package}
\alias{trans}
\docType{package}
\title{
\packageTitle{trans}
}
\description{
\packageDescription{trans}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{trans}
\packageIndices{trans}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{trans}

Maintainer: \packageMaintainer{trans}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
