% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{trans_add__generic}
\alias{trans_add__generic}
\title{Adds a generic column based on an arbitrary mutation function \lifecycle{experimental}}
\usage{
trans_add__generic(dat, .col_source, .col_target, .fun, .value,
  .verbose = confx::conf_get("tracing/verbose", "config.yml"))
}
\arguments{
\item{dat}{[\link{tbl}] A \code{tibble} containing the data which should be transformed}

\item{.col_source}{[\code{\link[=character]{character()}}] Name of the source column that the
mutation function uses. Could also be a \code{\link{name}} and soon also
\code{\link{quosures}}.}

\item{.col_target}{[\code{\link[=character]{character()}}] Name of the target column that the output
of the mutation function is assigned to. Could also be a \code{\link{name}} and soon
also \code{\link{quosures}}.}

\item{.fun}{[\code{\link[=function]{function()}}] The mutation function to be used}

\item{.value}{[\code{\link[=ANY]{ANY()}}] The assignment value in case no function is provided}
}
\value{
[\link{tbl}] Tibble with additional column
}
\description{
Adds a generic column based on an arbitrary mutation function.
}
