% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{trans_bind_cols_auto}
\alias{trans_bind_cols_auto}
\title{Bind columns with automatic column handling \lifecycle{experimental}}
\usage{
trans_bind_cols_auto(x, y)
}
\arguments{
\item{x}{[\code{\link[=tibble]{tibble()}}]}

\item{y}{[\code{\link[=tibble]{tibble()}}]}
}
\value{
[\code{\link[=tibble]{tibble()}}]
}
\description{
TODO-20191112-2: add doc
}
