% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{trans_left_join_auto}
\alias{trans_left_join_auto}
\title{Left join with automatic column handling \lifecycle{experimental}}
\usage{
trans_left_join_auto(x, y, by = character())
}
\arguments{
\item{x}{[\code{\link[=tibble]{tibble()}}]}

\item{y}{[\code{\link[=tibble]{tibble()}}]}

\item{by}{[\code{\link[=character]{character()}}]}
}
\value{
[\code{\link[=tibble]{tibble()}}]
}
\description{
TODO-20191112-1: add doc
}
